change to double

+++ fortscale/fortscale-collection/src/main/java/fortscale/collection/jobs/activity/UserActivityBaseHandler.java
@@ -80,7 +80,7 @@ public abstract class UserActivityBaseHandler implements UserActivityHandler {
         int actualUserChunkSize = Math.min(MONGO_READ_WRITE_BULK_SIZE, numberOfUsers);
         int numOfHandledUsers;
 
-        Map<String, Integer> additionalActivityHistogram = new HashMap<>();
+        Map<String, Double> additionalActivityHistogram = new HashMap<>();
 
         long currBucketStartTime = firstBucketStartTime;
         long currBucketEndTime = firstBucketEndTime;
@@ -142,7 +142,7 @@ public abstract class UserActivityBaseHandler implements UserActivityHandler {
 
     }
 
-    protected Map<String, Integer> updateAdditionalActivitySpecificHistograms(Map<String, UserActivityDocument> userActivityMap) {
+    protected Map<String, Double> updateAdditionalActivitySpecificHistograms(Map<String, UserActivityDocument> userActivityMap) {
         return Collections.emptyMap();
     }
 
@@ -280,7 +280,7 @@ public abstract class UserActivityBaseHandler implements UserActivityHandler {
     }
     private void updateActivitySpecificHistogram(Map<String, UserActivityDocument> userActivityMap, FeatureBucket featureBucket, String contextId) {
         UserActivityDocument userActivityDocument = userActivityMap.get(contextId);
-        Map<String, Integer> histogramOfUser = userActivityDocument.getHistogram();
+        Map<String, Double> histogramOfUser = userActivityDocument.getHistogram();
 
         final Map<String, Feature> aggregatedFeatures = featureBucket.getAggregatedFeatures();
         final List<String> histogramFeatureNames = getRelevantAggregatedFeaturesFieldsNames();
@@ -289,8 +289,8 @@ public abstract class UserActivityBaseHandler implements UserActivityHandler {
             final GenericHistogram featureAsHistogram = convertFeatureToHistogram(featureValue, histogramFeatureName);
             Map<String, Double> bucketHistogram = featureAsHistogram.getHistogramMap();
             for (Map.Entry<String, Double> entry : bucketHistogram.entrySet()) {
-                int oldValue = histogramOfUser.get(entry.getKey()) != null ? histogramOfUser.get(entry.getKey()) : 0;
-                int newValue = entry.getValue().intValue();
+                double oldValue = histogramOfUser.get(entry.getKey()) != null ? histogramOfUser.get(entry.getKey()) : 0;
+                double newValue = entry.getValue();
                 histogramOfUser.put(entry.getKey(), oldValue + newValue);
             }
         }
@@ -323,6 +323,6 @@ public abstract class UserActivityBaseHandler implements UserActivityHandler {
      * @param currBucketEndTime the end time of the document
      * @param additionalActivityHistogram the histogram of the document
      */
-    protected abstract void updateAdditionalActivitySpecificDocumentInDatabase(List<String> dataSources, long currBucketStartTime, long currBucketEndTime, Map<String, Integer> additionalActivityHistogram);
+    protected abstract void updateAdditionalActivitySpecificDocumentInDatabase(List<String> dataSources, long currBucketStartTime, long currBucketEndTime, Map<String, Double> additionalActivityHistogram);
 
 }
