+++ fortscale/fortscale-domain/src/main/java/fortscale/domain/core/activities/UserActivityNetworkAuthenticationDocument.java
@@ -31,16 +31,16 @@ public class UserActivityNetworkAuthenticationDocument extends UserActivityDocum
 	}
 
 	@Override
-	public Map<String, Integer> getHistogram() {
+	public Map<String, Double> getHistogram() {
 		return getAuthentications().getAuthenticationsHistogram();
 	}
 
 	public static class Authentications {
-		private Map<String, Integer> authenticationsHistogram = new HashMap<>();
+		private Map<String, Double> authenticationsHistogram = new HashMap<>();
 
 
 		@Field(AUTHENTICATIONS_HISTOGRAM_FIELD_NAME)
-		public Map<String, Integer> getAuthenticationsHistogram() {
+		public Map<String, Double> getAuthenticationsHistogram() {
 			return authenticationsHistogram;
 		}
 	}
